Replace some fixes of 2^31 limit with exception.
We can easily address these later if needed, but if we're unlikely
to encounter the issue, it's probably better to keep the limitation and
not incur any performance penalty (or have to mitigate them using
alternate implementations).

+++ engine/src/main/java/nl/inl/blacklab/resultproperty/HitProperty.java
@@ -190,8 +190,8 @@ public abstract class HitProperty implements ResultProperty<Hit>, LongComparator
     
             // Unless the client sets different context indices, assume we got the ones we wanted in the correct order
             if (contexts != null) {
-                long numberOfAnnotations = contexts.annotations().size();
-                this.contextIndices = new IntArrayList((int)numberOfAnnotations);
+                int numberOfAnnotations = contexts.annotations().size();
+                this.contextIndices = new IntArrayList(numberOfAnnotations);
                 for (int i = 0; i < numberOfAnnotations; ++i) this.contextIndices.add(i, i);
             }
         }
